requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-filter==23.5
python-decouple==3.8
waitress==3.0.0
hypothesis==6.92.1
requests==2.31.0